backend/requirements.txt
Django>=3.2
djangorestframework>=3.12
djoser>=2.1.0
django-cors-headers>=3.5.0
gunicorn==20.1.0
coverage==5.5